attempt for better reporting of the generateAPIChanges.sh
Attempted to add a new features that looks for changes or not to the
API and decide to do a few tasks. However, I could not find any answers
at this time to get this code to work inside a script: grep -q ‘No API
changes found.” report.txt
This work just fine outside the script but not within it. Decided to
render that info to the shell and still open the HTML generated file
regardless if it exists or not.