version target tweaks
Added notices to the generateAPIChanges.sh script notifying the user to
iterate above the current (end) version so that version is including in
the report. I also updated the fetch/update routines and included a
fetch/update for Titanium_mobile_windows, a check for the js-yml NPM
module, and the argument to include Titanium_mobile_windows into the
report.